requirements.txt
langchain
langchain_community
pinecone
langchain_google_genai
streamlit
python-dotenv